+++ glesgae/gae_point.h
@@ -24,7 +24,7 @@ typedef struct gae_point_4d_s {
 /* float api */
 
 /* work out float distance between two points, given as separate floats */
-float gae_point_distancef(float x1, float y1, float x2, float y2);
+double gae_point_distancef(float x1, float y1, float x2, float y2);
 
 /* work out float distance between two points, given as separate floats, leaving it squared */
 float gae_point_distancef2(float x1, float y1, float x2, float y2);
@@ -38,7 +38,7 @@ int gae_point2d_in_rectangle(gae_point_2d_t point, gae_rect_t rect);
 gae_point_2d_t gae_point2d_rotate(gae_point_2d_t* const origin, float radians, gae_point_2d_t* const p);
 
 /* returns angle between two points */
-float gae_point2d_angle_between(gae_point_2d_t* const a, gae_point_2d_t* const b);
+double gae_point2d_angle_between(gae_point_2d_t* const a, gae_point_2d_t* const b);
 
 /* scales the given point with the given scaler */
 gae_point_2d_t* gae_point2d_scale(gae_point_2d_t* const point, float scaler);
